LayoutTests/imported/w3c/web-platform-tests/css/css-values/getComputedStyle-border-radius-003-expected.txt


FAIL testing border-radius shorthand assert_equals: expected "calc(1% + 1px) calc(2% + 2px) calc(3% + 3px) calc(4% + 4px) / calc(5% + 5px) calc(6% + 6px) calc(7% + 7px) calc(8% + 8px)" but got "calc(1px + 1%) calc(2px + 2%) calc(3px + 3%) calc(4px + 4%) / calc(5px + 5%) calc(6px + 6%) calc(7px + 7%) calc(8px + 8%)"
